backend/requirements.txt
altair
annotated-doc
annotated-types
anyio
astrapy
asttokens
attrs
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cycler
decorator
deprecation
distro
dnspython
executing
fastapi
filelock
filetype
fonttools
fsspec
gitdb
GitPython
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.69.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
grpcio
grpcio-status
h11
h2
hf-xet
hpack
httpcore
httplib2
httpx
httpx-sse
huggingface_hub
hyperframe
idna
ipython
ipython_pygments_lexers
jedi
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
langchain
langchain-classic
langchain-astradb
langchain-core
langchain-google-genai
langchain-mistralai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mpmath
narwhals
networkx
nltk
numpy
opencv-python
orjson
ormsgpack
packaging
pandas
parso
pillow
prompt_toolkit
proto-plus
protobuf
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydeck
Pygments
pymongo
pyparsing
PyPDF2
python-dateutil
python-docx
python-dotenv
python-multipart
pytz
PyYAML
referencing
regex
requests
requests-toolbelt
rich
rpds-py
scikit-learn
scipy
shellingham
six
smmap
sniffio
stack-data
starlette
streamlit
sympy
tenacity
threadpoolctl
tokenizers
toml
torch
torchaudio
torchvision
tornado
tqdm
traitlets
typer
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uuid6
uuid_utils
uvicorn
watchdog
wcwidth
websockets
xxhash
zstandard
